chore(weave): extract conditions.py from calls_query_builder
Extract query condition processing into a dedicated conditions.py module.
This eliminates the circular dependency between calls_query_builder.py
and object_ref_query_builder.py.

Key change: object_ref_query_builder.py now imports process_query_to_conditions
from conditions.py at module level, replacing the fragile runtime deferred
import that was needed to work around the circular dependency.

Extracted: FilterToConditions, process_query_to_conditions,
_maybe_convert_datetime_operands, _extract_field_name
